3단계 - 로또(2등) (#2624)
* feat : LottoReward rank

* feat : LottoOutput bonusBall

* feat : LottoInput bonusNumber

* feat : LottoOutput matchBonusNumber

* refactor : LotoNumberSet

* refactor : Lotto rank

* refactor : LottoNumberSet sort toString

* feat : LottoWinner valid

* feat : LottoWinner winner

* test : LottoWinner rank

* refactor : LottoReward reward

* refactor : LottoController lottoResult

* refactor : LottoWinner rank

* refactor : Money int에서 long으로 변경

* refactor : cleanup code

* refactor : LottoService reward

* refactor : LottoController lottoResult

* style : 파라미터 네이밍 수정

+++ src/main/java/lotto/domain/Money.java
@@ -4,10 +4,10 @@ import java.util.Objects;
 
 public class Money {
 
-    private int money;
+    private long money;
     private static String MONEY_NEGATIVE_EXCEPTION = "Money는 음수가 될 수 없습니다.";
 
-    public Money(int money) {
+    public Money(long money) {
         if (money < 0) {
             throw new IllegalArgumentException(MONEY_NEGATIVE_EXCEPTION);
         }
@@ -18,7 +18,7 @@ public class Money {
         return (double) this.money / money.money();
     }
 
-    public int money() {
+    public long money() {
         return money;
     }
